pyproject.toml
[tool.poetry]
name = "nonebot_plugin_pjsk"
version = "0.1.0"
description = "PJSK表情包for Nonebot2"
authors = ["Agnes_Digital <Z735803792@163.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Agnes4m/nonebot_plugin_pjsk"
repository = "https://github.com/Agnes4m/nonebot_plugin_pjsk"
keywords = ["pjsk", "nonebot2", "plugin"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",

]
include = [
    "LICENSE","README.md"
]
 
[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = ">=2.1.3"
pillow = ">=8.4.0"


[tool.poetry.dev-dependencies]

[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
